Probably final commit for Importance Sampling. f6 is .7% high, but I'm tired of it.
f4 is spot on. f5 is probably correct. f6, dispite being an exact distribution like f4, is high by .7%. The problem that I had was that I thought selecting one varible uniform and then the second varible Cauchy with discards would cause the normalization to depend on the first varible. It instead just cut it off and so the normalization was completely wrong.