Last update to Importance sampling.
The problem that I'm running into with function 6 is that it is the sum of two functions and I'm modeling it with a coin flip between two distributions. Each distribution models their function exactly as in function 4 but the coin flip needs to be weighted in proportion to the source function. In otherwords, instead of doing what I've done here, each function should be integrated individually and then added together rather than added and then integrated with two distributions. If the answer was unknown, then choosing the weighting a priori becomes impossible.