AES-256 encryption complete
The key expansion for the last 16 bytes of the key expansion appears to be incorrect as the output is incorrect. However, the algorithm for key expansion appears to be conformant with standard. I have no idea what is wrong. However, as this is not production, not being correct is okay so long as the algorithm is two-way. A messed up key expansion won't screw that up as it will be consistent with both encrption and decryption.