Update to Server and Client
Both now make a message with an AES key and a hash of the RSA key exchange.
Client is ready to send the raw bytes. Server needs a socket in the Client object. Once the raw bytes are exchanged, I'll get the message encyrpted in RSA and then converted to extended Hamming(512, 496).