Room reverb made
Probably some fiddeling with the FFT to get it to work or see what state it may be in.
Simple room reverb is added in. It calculates the impulse response of a room using image speakers. It then has the convolution reverb actually calculuate the reverb. It assumes the reference source is at 1 meter so that any speaker closer than 1 meter is amplified and any speaker further than 1 meter is attenuated by the distance